# perf(ext/event): replace ReflectHas with object lookup