--- a/ext/web/02_event.js
+++ b/ext/web/02_event.js
@@ -31,7 +31,6 @@ const {
   ObjectGetOwnPropertyDescriptor,
   ObjectPrototypeIsPrototypeOf,
   ReflectDefineProperty,
-  ReflectHas,
   SafeArrayIterator,
   SafeMap,
   StringPrototypeStartsWith,
@@ -113,7 +112,7 @@ function setStopImmediatePropagation(
 function hasRelatedTarget(
   event,
 ) {
-  return ReflectHas(event, "relatedTarget");
+  return !!event.relatedTarget;
 }
 
 const isTrusted = ObjectGetOwnPropertyDescriptor({
@@ -501,9 +500,12 @@ function isShadowRoot(nodeImpl) {
 }
 
 function isSlottable(
-  nodeImpl,
+  /* nodeImpl, */
 ) {
-  return Boolean(isNode(nodeImpl) && ReflectHas(nodeImpl, "assignedSlot"));
+  // TODO(marcosc90) currently there aren't any slottables nodes
+  // https://dom.spec.whatwg.org/#concept-slotable
+  // return isNode(nodeImpl) && ReflectHas(nodeImpl, "assignedSlot");
+  return false;
 }
 
 // DOM Logic functions
@@ -972,7 +974,7 @@ class EventTarget {
 
     const { listeners } = self[eventTargetData];
 
-    if (!(ReflectHas(listeners, type))) {
+    if (!listeners[type]) {
       listeners[type] = [];
     }
 
@@ -1020,7 +1022,7 @@ class EventTarget {
     );
 
     const { listeners } = self[eventTargetData];
-    if (callback !== null && ReflectHas(listeners, type)) {
+    if (callback !== null && listeners[type]) {
       listeners[type] = ArrayPrototypeFilter(
         listeners[type],
         (listener) => listener.callback !== callback,
@@ -1069,7 +1071,7 @@ class EventTarget {
     }
 
     const { listeners } = self[eventTargetData];
-    if (!ReflectHas(listeners, event.type)) {
+    if (!listeners[event.type]) {
       setTarget(event, this);
       return true;
     }