@mmastrac 2023-08-17
Could we also add another bench call with an `onfoo` handler?