@marcosc90 2023-08-17
> Could we also add another bench call with an `onfoo` handler?

You mean with `handleEvent`? or what exactly

```
  const listener = {
    handleEvent() { /* */ },
  };
```