@bartlomieju 2023-08-17
@marcosc90 I believe Matt meant for a benchmark like that:
```
tg.onfoo = () => {};
Deno.bench("event dispatch ", () => {
  tg.dispatchEvent(ev);
});
```