@mmastrac 2023-08-17
@bartlomieju @marcosc90 

Correct, though it requires a call to `defineEventHandler` to set it up:

(from the websocket code)

```ts
defineEventHandler(WebSocket.prototype, "message");
defineEventHandler(WebSocket.prototype, "error");
defineEventHandler(WebSocket.prototype, "close");
defineEventHandler(WebSocket.prototype, "open");
```