@mmastrac 2023-08-17
`defineEventHandler` might be tricky to access, but I think you can dispatch an abort event to an AbortController's `onabort` signal like so to test this path:

```
let s = new AbortController().signal
s.onabort = function() { /* ... */ }
```