@marcosc90 2023-08-17
@mmastrac I exposed `defineEventHandler` on the global scope so I can bench it

```js
class WS extends EventTarget{}
defineEventHandler(WS.prototype, "message");

const ws = new WS();
ws.onmessage = (e) => {};

const event = new MessageEvent("message", {
  data: 'foo',
  origin: 'localhost',
});

Deno.bench("event dispatch onmessage", () => {
 ws.dispatchEvent(event);
});
```

**this PR**
```
cpu: 13th Gen Intel(R) Core(TM) i9-13900H
runtime: deno 1.36.1 (x86_64-unknown-linux-gnu)

benchmark                     time (avg)        iter/s             (min … max)       p75       p99      p995
------------------------------------------------------------------------------ -----------------------------
event dispatch                80.23 ns/iter  12,464,844.2  (73.97 ns … 121.41 ns)  81.45 ns  87.38 ns  92.71 ns
event dispatch onmessage      79.63 ns/iter  12,558,573.3   (73.55 ns … 87.74 ns)  80.96 ns  84.11 ns  85.41 ns
```

**main**
```
cpu: 13th Gen Intel(R) Core(TM) i9-13900H
runtime: deno 1.36.1 (x86_64-unknown-linux-gnu)

benchmark                     time (avg)        iter/s             (min … max)       p75       p99      p995
------------------------------------------------------------------------------ -----------------------------
event dispatch               102.54 ns/iter   9,752,364.8  (96.53 ns … 249.32 ns) 103.38 ns  116.9 ns 132.43 ns
event dispatch onmessage     106.07 ns/iter   9,427,655.9 (100.12 ns … 120.78 ns) 107.28 ns 111.81 ns 112.02 ns
```