@marcosc90 2023-08-18
> LGTM, thanks Marcos! Once we run our internal benches we'll let you know how that impact real world benchmarks.

I'm making improvements on `addEventListener` now